backend/requirements.txt
fastapi
uvicorn
google-api-python-client
requests
python-dotenv
transformers
torch
tweepy
apify-client